Use more fine grained storage access permissions
In order to run the notebook steps without errors, we need:
* `roles/storage.objectCreator`
* `roles/storage.objectViewer`
* `storage.buckets.get` <-- this isn't really necessary but the libs we
  are calling will try to see if the bucket we manually create exists
  and issue a noisy error if not

The only appropriate existing role that included `storage.buckets.get`
was the storage admin role which is overkill if we don't want the
account to actually be managing buckets. So instead we create a customer
role with exactly what we need.